backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.7
groq==0.4.1
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
sentence-transformers==2.2.2